+++ lib/src/draggable_data.dart
@@ -1,0 +1,9 @@
+import 'package:tabbed_view/src/tab_data.dart';
+import 'package:tabbed_view/src/tabbed_view_controller.dart';
+
+class DraggableData {
+  DraggableData(this.controller, this.tabData);
+
+  final TabbedViewController controller;
+  final TabData tabData;
+}