+++ lib/src/internal/tabs_area/drop_tab_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
+import 'package:tabbed_view/src/draggable_data.dart';
 import 'package:tabbed_view/src/internal/tabbed_view_provider.dart';
-import 'package:tabbed_view/src/tab_data.dart';
 
 @internal
 class DropTabWidget extends StatefulWidget {
@@ -14,6 +14,8 @@ class DropTabWidget extends StatefulWidget {
   final TabbedViewProvider provider;
   final Widget child;
   final int newIndex;
+
+  static const double dropWidth = 8;
 
   @override
   State<StatefulWidget> createState() => DropTabWidgetState();
@@ -32,7 +34,7 @@ class DropTabWidgetState extends State<DropTabWidget> {
 
   @override
   Widget build(BuildContext context) {
-    return DragTarget<TabData>(
+    return DragTarget<DraggableData>(
       builder: (
         BuildContext context,
         List<dynamic> accepted,
@@ -45,21 +47,52 @@ class DropTabWidgetState extends State<DropTabWidget> {
         return widget.child;
       },
       onMove: (details) {
-        if (_over == false) {
+        if (_over == false && _canDrop(details.data)) {
           setState(() {
             _over = true;
           });
         }
       },
       onLeave: (data) {
-        setState(() {
-          _over = false;
-        });
+        if (_over) {
+          setState(() {
+            _over = false;
+          });
+        }
       },
-      onAccept: (TabData tabData) {
-        widget.provider.controller.reorderTab(tabData.index, widget.newIndex);
+      onWillAccept: (data) {
+        if (data != null) {
+          return _canDrop(data);
+        }
+        return false;
+      },
+      onAccept: (DraggableData data) {
+        if (widget.provider.onBeforeDropAccept != null) {
+          if (widget.provider.onBeforeDropAccept!(
+                  data, widget.provider.controller) ==
+              false) {
+            setState(() {
+              _over = false;
+            });
+            return;
+          }
+        }
+        if (widget.provider.controller == data.controller) {
+          widget.provider.controller
+              .reorderTab(data.tabData.index, widget.newIndex);
+        } else {
+          data.controller.removeTab(data.tabData.index);
+          widget.provider.controller.insertTab(widget.newIndex, data.tabData);
+        }
       },
     );
+  }
+
+  bool _canDrop(DraggableData source) {
+    if (widget.provider.canDrop == null) {
+      return true;
+    }
+    return widget.provider.canDrop!(source, widget.provider.controller);
   }
 }
 
@@ -67,9 +100,10 @@ class _CustomPainter extends CustomPainter {
   @override
   void paint(Canvas canvas, Size size) {
     Paint paint = Paint()
-      ..color = Colors.black
+      ..color = Colors.black.withOpacity(.7)
       ..style = PaintingStyle.fill;
-    canvas.drawRect(Rect.fromLTWH(0, 0, 3, size.height), paint);
+    canvas.drawRect(
+        Rect.fromLTWH(0, 0, DropTabWidget.dropWidth, size.height), paint);
   }
 
   @override