+++ lib/src/tabs_area.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:tabbed_view/src/flow_layout.dart';
+import 'package:tabbed_view/src/internal/tabs_area/drop_tab_widget.dart';
+import 'package:tabbed_view/src/internal/tabs_area/hidden_tabs.dart';
 import 'package:tabbed_view/src/internal/tabbed_view_provider.dart';
+import 'package:tabbed_view/src/internal/tabs_area/tabs_area_corner.dart';
 import 'package:tabbed_view/src/tab_button.dart';
 import 'package:tabbed_view/src/tab_button_widget.dart';
 import 'package:tabbed_view/src/tab_data.dart';
@@ -28,7 +30,7 @@ class TabsArea extends StatefulWidget {
 class _TabsAreaState extends State<TabsArea> {
   int? _highlightedIndex;
 
-  final HiddenTabs hiddenTabs = HiddenTabs();
+  final HiddenTabs _hiddenTabs = HiddenTabs();
 
   @override
   Widget build(BuildContext context) {
@@ -44,11 +46,14 @@ class _TabsAreaState extends State<TabsArea> {
           provider: widget.provider,
           updateHighlightedIndex: _updateHighlightedIndex));
     }
+
+    children.add(
+        TabsAreaCorner(provider: widget.provider, hiddenTabs: _hiddenTabs));
+
     Widget tabsAreaLayout = TabsAreaLayout(
         children: children,
-        buttonsAreaBuilder: _buttonsAreaBuilder,
         theme: theme,
-        hiddenTabs: hiddenTabs,
+        hiddenTabs: _hiddenTabs,
         selectedTabIndex: controller.selectedIndex);
     tabsAreaLayout = ClipRect(child: tabsAreaLayout);
 
@@ -58,86 +63,6 @@ class _TabsAreaState extends State<TabsArea> {
           color: tabsAreaTheme.color, border: tabsAreaTheme.border);
     }
     return Container(child: tabsAreaLayout, decoration: decoration);
-  }
-
-  /// Area for buttons like the hidden tabs menu button.
-  ///
-  /// Even if there are no visible buttons, an empty container must be created.
-  Widget _buttonsAreaBuilder(BuildContext context) {
-    TabbedViewThemeData theme = TabbedViewTheme.of(context);
-    TabsAreaThemeData tabsAreaTheme = theme.tabsArea;
-    Widget buttonsArea;
-
-    List<TabButton> buttons = [];
-    if (widget.provider.tabsAreaButtonsBuilder != null) {
-      buttons = widget.provider.tabsAreaButtonsBuilder!(
-          context, widget.provider.controller.tabs.length);
-    }
-
-    if (hiddenTabs.hasHiddenTabs) {
-      TabButton hiddenTabsMenuButton = TabButton(
-          icon: tabsAreaTheme.menuIcon, menuBuilder: _hiddenTabsMenuBuilder);
-      buttons.insert(0, hiddenTabsMenuButton);
-    }
-
-    if (buttons.isNotEmpty) {
-      List<Widget> children = [];
-      for (int i = 0; i < buttons.length; i++) {
-        EdgeInsets? padding;
-        if (i > 0 && tabsAreaTheme.buttonsGap > 0) {
-          padding = EdgeInsets.only(left: tabsAreaTheme.buttonsGap);
-        }
-        TabButton tabButton = buttons[i];
-        children.add(Container(
-            child: TabButtonWidget(
-                provider: widget.provider,
-                button: tabButton,
-                enabled: true,
-                normalColor: tabsAreaTheme.normalButtonColor,
-                hoverColor: tabsAreaTheme.hoverButtonColor,
-                disabledColor: tabsAreaTheme.disabledButtonColor,
-                normalBackground: tabsAreaTheme.normalButtonBackground,
-                hoverBackground: tabsAreaTheme.hoverButtonBackground,
-                disabledBackground: tabsAreaTheme.disabledButtonBackground,
-                iconSize: tabButton.iconSize != null
-                    ? tabButton.iconSize!
-                    : tabsAreaTheme.buttonIconSize,
-                themePadding: tabsAreaTheme.buttonPadding),
-            padding: padding));
-      }
-
-      buttonsArea = FlowLayout(children: children, firstChildFlex: false);
-
-      EdgeInsetsGeometry? margin;
-      if (tabsAreaTheme.buttonsOffset > 0) {
-        margin = EdgeInsets.only(left: tabsAreaTheme.buttonsOffset);
-      }
-      if (tabsAreaTheme.buttonsAreaDecoration != null ||
-          tabsAreaTheme.buttonsAreaPadding != null ||
-          margin != null) {
-        buttonsArea = Container(
-            child: buttonsArea,
-            decoration: tabsAreaTheme.buttonsAreaDecoration,
-            padding: tabsAreaTheme.buttonsAreaPadding,
-            margin: margin);
-      }
-    } else {
-      buttonsArea = SizedBox(width: 0);
-    }
-    return buttonsArea;
-  }
-
-  /// Builder for hidden tabs menu.
-  List<TabbedViewMenuItem> _hiddenTabsMenuBuilder(BuildContext context) {
-    List<TabbedViewMenuItem> list = [];
-    hiddenTabs.indexes.sort();
-    for (int index in hiddenTabs.indexes) {
-      TabData tab = widget.provider.controller.tabs[index];
-      list.add(TabbedViewMenuItem(
-          text: tab.text,
-          onSelection: () => widget.provider.controller.selectedIndex = index));
-    }
-    return list;
   }
 
   /// Gets the status of the tab for a given index.