+++ lib/src/tabs_area_layout.dart
@@ -1,6 +1,10 @@
 import 'dart:math' as math;
+
 import 'package:flutter/rendering.dart';
 import 'package:flutter/widgets.dart';
+import 'package:tabbed_view/src/internal/tabs_area/hidden_tabs.dart';
+import 'package:tabbed_view/src/internal/tabs_area/tabs_area_layout_parent_data.dart';
+import 'package:tabbed_view/src/internal/tabs_area/visible_tabs.dart';
 import 'package:tabbed_view/src/theme/equal_heights.dart';
 import 'package:tabbed_view/src/theme/tabbed_view_theme_data.dart';
 import 'package:tabbed_view/src/theme/tabs_area_theme_data.dart';
@@ -11,22 +15,11 @@ import 'package:tabbed_view/src/theme/tabs_area_theme_data.dart';
 class TabsAreaLayout extends MultiChildRenderObjectWidget {
   TabsAreaLayout(
       {Key? key,
-      required WidgetBuilder buttonsAreaBuilder,
       required List<Widget> children,
       required this.theme,
       required this.hiddenTabs,
       required this.selectedTabIndex})
-      : super(
-          key: key,
-          children: [
-            ...children,
-            LayoutBuilder(
-              builder: (context, constraints) {
-                return buttonsAreaBuilder(context);
-              },
-            ),
-          ],
-        );
+      : super(key: key, children: children);
 
   final TabbedViewThemeData theme;
   final HiddenTabs hiddenTabs;
@@ -61,8 +54,8 @@ class _TabsAreaLayoutElement extends MultiChildRenderObjectElement {
   void debugVisitOnstageChildren(ElementVisitor visitor) {
     children.forEach((child) {
       if (child.renderObject != null) {
-        _TabsAreaLayoutParentData parentData =
-            child.renderObject!.parentData as _TabsAreaLayoutParentData;
+        TabsAreaLayoutParentData parentData =
+            child.renderObject!.parentData as TabsAreaLayoutParentData;
         if (parentData.visible) {
           visitor(child);
         }
@@ -71,169 +64,11 @@ class _TabsAreaLayoutElement extends MultiChildRenderObjectElement {
   }
 }
 
-/// Holds the hidden tab indexes.
-class HiddenTabs {
-  List<int> indexes = [];
-
-  bool get hasHiddenTabs => indexes.isNotEmpty;
-}
-
-/// Constraints with value for [TabsAreaLayout].
-class _TabsAreaButtonsBoxConstraints extends BoxConstraints {
-  _TabsAreaButtonsBoxConstraints({
-    required HiddenTabs hiddenTabs,
-    required BoxConstraints constraints,
-  })  : this.hasHiddenTabs = hiddenTabs.hasHiddenTabs,
-        super(
-            minWidth: constraints.minWidth,
-            maxWidth: constraints.maxWidth,
-            minHeight: constraints.minHeight,
-            maxHeight: constraints.maxHeight);
-
-  final bool hasHiddenTabs;
-
-  @override
-  bool operator ==(dynamic other) {
-    assert(debugAssertIsValid());
-    if (identical(this, other)) return true;
-    if (other is! _TabsAreaButtonsBoxConstraints) return false;
-    final _TabsAreaButtonsBoxConstraints typedOther = other;
-    assert(typedOther.debugAssertIsValid());
-    return hasHiddenTabs == typedOther.hasHiddenTabs &&
-        minWidth == typedOther.minWidth &&
-        maxWidth == typedOther.maxWidth &&
-        minHeight == typedOther.minHeight &&
-        maxHeight == typedOther.maxHeight;
-  }
-
-  @override
-  int get hashCode {
-    assert(debugAssertIsValid());
-    return Object.hash(minWidth, maxWidth, minHeight, maxHeight, hasHiddenTabs);
-  }
-}
-
-/// Parent data for [_TabsAreaLayoutRenderBox] class.
-class _TabsAreaLayoutParentData extends ContainerBoxParentData<RenderBox> {
-  bool visible = false;
-  bool selected = false;
-
-  double leftBorderHeight = 0;
-  double rightBorderHeight = 0;
-
-  /// Resets all values.
-  reset() {
-    visible = false;
-    selected = false;
-
-    leftBorderHeight = 0;
-    rightBorderHeight = 0;
-  }
-}
-
-/// Utility class to find out which tabs may be visible.
-/// Calculates the required width and sets the offset value.
-class _VisibleTabs {
-  _VisibleTabs(this.tabsAreaTheme);
-
-  final TabsAreaThemeData tabsAreaTheme;
-
-  List<RenderBox> _tabs = [];
-
-  void add(RenderBox tab) {
-    _tabs.add(tab);
-  }
-
-  int get length => _tabs.length;
-
-  RenderBox get(int index) {
-    return _tabs[index];
-  }
-
-  /// Layouts the single selected tab.
-  layoutSingleTab(double maxWidth, double maxHeight, Size tabsAreaButtonsSize) {
-    if (_tabs.length == 1) {
-      double availableWidth = maxWidth -
-          tabsAreaButtonsSize.width -
-          tabsAreaTheme.initialGap -
-          tabsAreaTheme.minimalFinalGap;
-
-      if (availableWidth > 0) {
-        RenderBox tab = _tabs.first;
-        if (tab.size.width > availableWidth) {
-          final BoxConstraints childConstraints =
-              BoxConstraints.loose(Size(availableWidth, maxHeight));
-          tab.layout(childConstraints, parentUsesSize: true);
-        }
-      }
-    }
-  }
-
-  /// Updates the offset given the tab width, initial offset and gap values.
-  void updateOffsets(Size tabsAreaButtonsSize) {
-    double offset = tabsAreaTheme.initialGap;
-    for (int i = 0; i < _tabs.length; i++) {
-      RenderBox tab = _tabs[i];
-      _TabsAreaLayoutParentData tabParentData = tab.tabsAreaLayoutParentData();
-
-      tabParentData.offset = Offset(offset, tabParentData.offset.dy);
-
-      if (i < _tabs.length - 1) {
-        offset += tab.size.width + tabsAreaTheme.middleGap;
-      } else {
-        offset += tab.size.width + tabsAreaTheme.minimalFinalGap;
-      }
-    }
-  }
-
-  /// Calculates the required width for the tab. Includes the gap values
-  /// on the right.
-  double requiredTabWidth(int index) {
-    RenderBox tab = _tabs[index];
-    if (index < _tabs.length - 1) {
-      return tab.size.width + tabsAreaTheme.middleGap;
-    }
-    return tab.size.width + tabsAreaTheme.minimalFinalGap;
-  }
-
-  /// Calculates the required width for all tabs.
-  double requiredTotalWidth() {
-    double width = 0;
-    for (int i = 0; i < _tabs.length; i++) {
-      width += requiredTabWidth(i);
-    }
-    return width;
-  }
-
-  /// Removes the last non-selected tab. Returns the removed tab index.
-  int? removeLastNonSelected() {
-    int index = _tabs.length - 1;
-    while (index >= 0) {
-      RenderBox tab = _tabs[index];
-      _TabsAreaLayoutParentData parentData =
-          tab.parentData as _TabsAreaLayoutParentData;
-      if (parentData.selected == false) {
-        _tabs.removeAt(index);
-        return index;
-      }
-      index--;
-    }
-    return null;
-  }
-
-  /// Removes the first tab.
-  void removeFirst() {
-    if (_tabs.length > 0) {
-      _tabs.removeAt(0);
-    }
-  }
-}
-
 /// The [TabsAreaLayout] render box.
 class _TabsAreaLayoutRenderBox extends RenderBox
     with
-        ContainerRenderObjectMixin<RenderBox, _TabsAreaLayoutParentData>,
-        RenderBoxContainerDefaultsMixin<RenderBox, _TabsAreaLayoutParentData> {
+        ContainerRenderObjectMixin<RenderBox, TabsAreaLayoutParentData>,
+        RenderBoxContainerDefaultsMixin<RenderBox, TabsAreaLayoutParentData> {
   _TabsAreaLayoutRenderBox(
       TabbedViewThemeData theme, HiddenTabs hiddenTabs, int? selectedTabIndex)
       : this._tabsAreaTheme = theme.tabsArea,
@@ -243,6 +78,8 @@ class _TabsAreaLayoutRenderBox extends RenderBox
   int? _selectedTabIndex;
 
   int? get selectedTabIndex => _selectedTabIndex;
+
+  late RenderBox _corner;
 
   set selectedTabIndex(int? value) {
     if (_selectedTabIndex != value) {
@@ -269,70 +106,66 @@ class _TabsAreaLayoutRenderBox extends RenderBox
   set hiddenTabs(HiddenTabs value) {
     if (_hiddenTabs != value) {
       _hiddenTabs = value;
-      markNeedsLayout();
     }
   }
 
   @override
   void setupParentData(RenderBox child) {
-    if (child.parentData is! _TabsAreaLayoutParentData) {
-      child.parentData = _TabsAreaLayoutParentData();
+    if (child.parentData is! TabsAreaLayoutParentData) {
+      child.parentData = TabsAreaLayoutParentData();
     }
   }
 
   @override
   void performLayout() {
+    final BoxConstraints childConstraints =
+        BoxConstraints.loose(Size(double.infinity, constraints.maxHeight));
+
     List<RenderBox> children = [];
+    visitChildren((child) {
+      children.add(child as RenderBox);
+    });
+
+    // There will always be at least 1 child (corner area).
+    _corner = children.removeLast();
 
     // Defines the biggest height to avoid displacement of tabs when
     // changing visibility.
     double height = 0;
 
-    BoxConstraints childConstraints =
-        BoxConstraints.loose(Size(double.infinity, constraints.maxHeight));
-
-    hiddenTabs.indexes.clear();
+    _corner.layout(childConstraints, parentUsesSize: true);
+    final double minCornerAreaWidth = _corner.size.width;
+    height = math.max(height, _corner.size.height);
 
     // layout all (tabs + tabs area buttons)
-    _VisibleTabs visibleTabs = _VisibleTabs(tabsAreaTheme);
+    VisibleTabs visibleTabs = VisibleTabs(tabsAreaTheme);
 
-    // There will always be at least 1 child (tabs area buttons).
-    visitChildren((child) {
-      final _TabsAreaLayoutParentData parentData =
+    for (int i = 0; i < children.length; i++) {
+      final RenderBox child = children[i];
+      final TabsAreaLayoutParentData parentData =
           child.tabsAreaLayoutParentData();
       parentData.reset();
       parentData.selected =
-          selectedTabIndex != null && selectedTabIndex == children.length
-              ? true
-              : false;
-      children.add(child as RenderBox);
+          selectedTabIndex != null && selectedTabIndex == i ? true : false;
       child.layout(childConstraints, parentUsesSize: true);
       height = math.max(height, child.size.height);
+      visibleTabs.add(child);
+    }
 
-      if (child != lastChild) {
-        // ignoring tabs area buttons
-        visibleTabs.add(child);
-      }
-    });
     if (tabsAreaTheme.gapBottomBorder.style == BorderStyle.solid &&
         tabsAreaTheme.gapBottomBorder.width > 0) {
       height = math.max(height, tabsAreaTheme.gapBottomBorder.width);
     }
 
-    childConstraints = BoxConstraints.loose(Size(constraints.maxWidth, height));
-
-    // lastChild is the tabs area buttons and will always exist and be visible.
-    // It can be an empty SizedBox, a single button or a Container with buttons.
-    RenderBox tabsAreaButtons = lastChild!;
-
     double availableWidth = math.max(
-        constraints.maxWidth -
-            tabsAreaTheme.initialGap -
-            tabsAreaButtons.size.width,
+        constraints.maxWidth - tabsAreaTheme.initialGap - _corner.size.width,
         0);
 
     visibleTabs.layoutSingleTab(
-        constraints.maxWidth, height, tabsAreaButtons.size);
+        constraints.maxWidth, height, _corner.size.width);
+
+    List<int> hiddenIndexes = [];
+
     while (visibleTabs.length > 0 &&
         visibleTabs.requiredTotalWidth() > availableWidth) {
       int? removedIndex;
@@ -343,51 +176,45 @@ class _TabsAreaLayoutRenderBox extends RenderBox
         removedIndex = visibleTabs.removeLastNonSelected();
       }
       if (removedIndex != null) {
-        hiddenTabs.indexes.add(removedIndex);
-
-        // It performs another layout because the component can be changed
-        // with hidden tabs.
-        tabsAreaButtons.layout(
-          _TabsAreaButtonsBoxConstraints(
-              hiddenTabs: hiddenTabs, constraints: childConstraints),
-          parentUsesSize: true,
-        );
-        availableWidth = math.max(
-            constraints.maxWidth -
-                tabsAreaTheme.initialGap -
-                tabsAreaButtons.size.width,
-            0);
+        hiddenIndexes.add(removedIndex);
       }
       visibleTabs.layoutSingleTab(
-          constraints.maxWidth, height, tabsAreaButtons.size);
+          constraints.maxWidth, height, _corner.size.width);
     }
-    visibleTabs.updateOffsets(tabsAreaButtons.size);
+
+    hiddenTabs.update(hiddenIndexes);
+
+    visibleTabs.updateOffsets();
+
+    _corner.layout(
+        BoxConstraints.tightFor(
+            width: math.max(
+                constraints.maxWidth - visibleTabs.maxX(), minCornerAreaWidth),
+            height: height),
+        parentUsesSize: true);
 
     List<RenderBox> visibleChildren = [];
     for (int i = 0; i < visibleTabs.length; i++) {
       RenderBox tab = visibleTabs.get(i);
-      final _TabsAreaLayoutParentData tabParentData =
+      final TabsAreaLayoutParentData tabParentData =
           tab.tabsAreaLayoutParentData();
       tabParentData.visible = true;
       visibleChildren.add(tab);
     }
 
-    if (tabsAreaButtons.size.width > 0 &&
-        tabsAreaButtons.size.width <=
-            constraints.maxWidth - tabsAreaTheme.initialGap) {
-      final _TabsAreaLayoutParentData tabsAreaButtonsParentData =
-          tabsAreaButtons.tabsAreaLayoutParentData();
-      tabsAreaButtonsParentData.visible = true;
-      // anchoring tabsAreaButtons to the right
-      tabsAreaButtonsParentData.offset =
-          Offset(constraints.maxWidth - tabsAreaButtons.size.width, 0);
-      visibleChildren.add(tabsAreaButtons);
-    }
+    final TabsAreaLayoutParentData cornerParentData =
+        _corner.tabsAreaLayoutParentData();
+    // anchoring corner to the right
+    cornerParentData.offset = Offset(constraints.maxWidth - _corner.size.width,
+        constraints.maxHeight - _corner.size.height);
+
+    cornerParentData.visible = true;
+    visibleChildren.add(_corner);
 
     if (tabsAreaTheme.equalHeights == EqualHeights.none) {
       // Aligning and fix max height on visible children.
       for (RenderBox tab in visibleChildren) {
-        final _TabsAreaLayoutParentData parentData =
+        final TabsAreaLayoutParentData parentData =
             tab.tabsAreaLayoutParentData();
         tab.layout(
             BoxConstraints(
@@ -403,6 +230,7 @@ class _TabsAreaLayoutRenderBox extends RenderBox
       if (tabsAreaTheme.equalHeights == EqualHeights.tabs) {
         int visibleCount = visibleChildren.length;
         if (hiddenTabs.hasHiddenTabs) {
+          // ignoring corner
           visibleCount--;
         }
         for (int i = 0; i < visibleCount; i++) {
@@ -410,16 +238,16 @@ class _TabsAreaLayoutRenderBox extends RenderBox
           tab.layout(
               BoxConstraints.tightFor(width: tab.size.width, height: height),
               parentUsesSize: true);
-          final _TabsAreaLayoutParentData parentData =
+          final TabsAreaLayoutParentData parentData =
               tab.tabsAreaLayoutParentData();
           parentData.offset = Offset(parentData.offset.dx, 0);
         }
         if (hiddenTabs.hasHiddenTabs) {
-          RenderBox tabsAreaButtons = visibleChildren.last;
-          final _TabsAreaLayoutParentData parentData =
-              tabsAreaButtons.tabsAreaLayoutParentData();
-          parentData.offset = Offset(
-              parentData.offset.dx, height - tabsAreaButtons.size.height);
+          RenderBox corner = visibleChildren.last;
+          final TabsAreaLayoutParentData parentData =
+              corner.tabsAreaLayoutParentData();
+          parentData.offset =
+              Offset(parentData.offset.dx, height - corner.size.height);
         }
       } else if (tabsAreaTheme.equalHeights == EqualHeights.all) {
         for (RenderBox child in visibleChildren) {
@@ -427,7 +255,7 @@ class _TabsAreaLayoutRenderBox extends RenderBox
               BoxConstraints.tightFor(width: child.size.width, height: height),
               parentUsesSize: true);
 
-          final _TabsAreaLayoutParentData parentData =
+          final TabsAreaLayoutParentData parentData =
               child.tabsAreaLayoutParentData();
           parentData.offset = Offset(parentData.offset.dx, 0);
         }
@@ -454,11 +282,11 @@ class _TabsAreaLayoutRenderBox extends RenderBox
   void paint(PaintingContext context, Offset offset) {
     List<RenderBox> visibleTabs = [];
     RenderBox? selectedTab;
-    _TabsAreaLayoutParentData? selectedTabParentData;
+    TabsAreaLayoutParentData? selectedTabParentData;
     visitVisibleChildren((RenderObject child) {
-      final _TabsAreaLayoutParentData childParentData =
+      final TabsAreaLayoutParentData childParentData =
           child.tabsAreaLayoutParentData();
-      if (child != lastChild) {
+      if (child != _corner) {
         if (childParentData.selected) {
           selectedTab = child as RenderBox;
           selectedTabParentData = childParentData;
@@ -498,7 +326,7 @@ class _TabsAreaLayoutRenderBox extends RenderBox
 
     for (int i = 0; i < visibleTabs.length; i++) {
       RenderBox tab = visibleTabs[i];
-      _TabsAreaLayoutParentData tabParentData = tab.tabsAreaLayoutParentData();
+      TabsAreaLayoutParentData tabParentData = tab.tabsAreaLayoutParentData();
       if (tabParentData.visible) {
         left += tab.size.width;
 
@@ -518,17 +346,16 @@ class _TabsAreaLayoutRenderBox extends RenderBox
       double lastX;
       if (visibleTabs.length > 0) {
         RenderBox lastTab = visibleTabs.last;
-        _TabsAreaLayoutParentData tabParentData =
+        TabsAreaLayoutParentData tabParentData =
             lastTab.tabsAreaLayoutParentData();
         lastX = offset.dx + tabParentData.offset.dx + lastTab.size.width;
       } else {
         lastX = offset.dx + tabsAreaTheme.initialGap;
       }
 
-      RenderBox? tabsAreaButtons = lastChild!;
       double lastGapWidth = offset.dx + size.width - lastX;
-      if (tabsAreaButtons.tabsAreaLayoutParentData().visible) {
-        lastGapWidth -= tabsAreaButtons.size.width;
+      if (_corner.tabsAreaLayoutParentData().visible) {
+        // lastGapWidth -= _corner.size.width;
       }
       if (lastGapWidth > 0) {
         canvas.drawRect(
@@ -544,7 +371,7 @@ class _TabsAreaLayoutRenderBox extends RenderBox
     List<RenderBox> visibleTabs = [];
     visitVisibleChildren((renderObject) {
       final RenderBox child = renderObject as RenderBox;
-      final _TabsAreaLayoutParentData childParentData =
+      final TabsAreaLayoutParentData childParentData =
           child.tabsAreaLayoutParentData();
       if (childParentData.selected) {
         visibleTabs.insert(0, child);
@@ -556,7 +383,7 @@ class _TabsAreaLayoutRenderBox extends RenderBox
     bool hitTest = false;
     visibleTabs.forEach((child) {
       if (!hitTest) {
-        final _TabsAreaLayoutParentData childParentData =
+        final TabsAreaLayoutParentData childParentData =
             child.tabsAreaLayoutParentData();
         hitTest = result.addWithPaintOffset(
           offset: childParentData.offset,
@@ -575,7 +402,7 @@ class _TabsAreaLayoutRenderBox extends RenderBox
 
 /// Utility extension to facilitate obtaining parent data.
 extension _TabsAreaLayoutParentDataGetter on RenderObject {
-  _TabsAreaLayoutParentData tabsAreaLayoutParentData() {
-    return this.parentData as _TabsAreaLayoutParentData;
+  TabsAreaLayoutParentData tabsAreaLayoutParentData() {
+    return this.parentData as TabsAreaLayoutParentData;
   }
 }