+++ lib/tabbed_view.dart
@@ -21,3 +21,6 @@ export 'src/icon_provider.dart';
 export 'src/tab_status.dart';
 export 'src/tab_leading_builder.dart';
 export 'src/draggable_config.dart';
+export 'src/draggable_data.dart';
+export 'src/typedefs/on_draggable_build.dart';
+export 'src/typedefs/can_drop.dart';